Unet/requirements.txt
numpy
matplotlib
pillow
pycocotools
torch
torchvision
swanlab
